0-linear.c

+++ 0x1E-search_algorithms/0-linear.c
@@ -10,13 +10,13 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i = 0;
+	int i = 0;
 
 	if (array == NULL || size == 0)
 		return (-1);
 
 
-	for (i = 0; i < size; i++)
+	for (i = 0; (size_t)i < size; i++)
 	{
 		printf("Value checked array[%i] = [%i]\n", i, array[i]);
 			if (array[i] == value)
